+++ SegundaEntregaLavieroCorreia/src/main/java/com/coderhouse/controller/ConcesionariaController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.coderhouse.dto.AsignacionGamaACocheDTO;
 import com.coderhouse.models.Concesionaria;
 import com.coderhouse.service.ConcesionariaService;
 
@@ -24,13 +25,16 @@ public class ConcesionariaController {
 	@Autowired
 	private ConcesionariaService concesionariaService;
 	
-	@GetMapping
+	@GetMapping(path = {"/",""})
 	public List<Concesionaria> getAllCoches() {
 		return concesionariaService.findAll();
 	};
 	
 	@GetMapping("/{cocheId}")
-	public ResponseEntity<Concesionaria> getCocheById(@PathVariable Long cocheId) {
+	public ResponseEntity<?> getCocheById(@PathVariable Long cocheId) {
+		if (cocheId == null) {
+			return ResponseEntity.badRequest().body("¡El Id del coche no debe ser nulo!");
+		}
 		try {
 			Concesionaria coche = concesionariaService.findById(cocheId);
 			return ResponseEntity.ok(coche); //200
@@ -52,8 +56,31 @@ public class ConcesionariaController {
 		}
 	} 
 	
+	@PostMapping("/asignarGama")
+	public ResponseEntity<?> asignarGamaACoche(@RequestBody AsignacionGamaACocheDTO dto){
+		
+		if (dto.getGamaId() == null || dto.getCocheId() == null) {
+			return ResponseEntity.badRequest().body("¡Ningun Id de gama o de coche debe ser nulo!");
+		}
+		try {
+			Concesionaria cocheActualizado = concesionariaService.asignarGamaACoche(
+					dto.getCocheId(), dto.getGamaId()
+					);
+			return ResponseEntity.ok(cocheActualizado);
+		} catch (IllegalStateException error) {
+			return ResponseEntity.status(HttpStatus.CONFLICT).body(error.getMessage()); //409
+		} catch(IllegalArgumentException error) {
+			return ResponseEntity.notFound().build(); //404	
+		} catch(Exception error) {
+			return ResponseEntity.internalServerError().build(); //500		
+		}
+	}
+	
 	@PutMapping("/{cocheId}")
-	public ResponseEntity<Concesionaria> updateCocheById(@PathVariable Long cocheId, @RequestBody Concesionaria cocheActualizado) {
+	public ResponseEntity<?> updateCocheById(@PathVariable Long cocheId, @RequestBody Concesionaria cocheActualizado) {
+		if (cocheId == null) {
+			return ResponseEntity.badRequest().body("¡El Id del coche no debe ser nulo!");
+		}
 		try {
 			Concesionaria coche = concesionariaService.update(cocheId, cocheActualizado);
 			return ResponseEntity.ok(coche); //200
@@ -65,7 +92,10 @@ public class ConcesionariaController {
 	}
 	
 	@DeleteMapping("/{cocheId}")
-	public ResponseEntity<Void> deleteCoche(@PathVariable Long cocheId) {
+	public ResponseEntity<?> deleteCoche(@PathVariable Long cocheId) {
+		if (cocheId == null) {
+			return ResponseEntity.badRequest().body("¡El Id del coche no debe ser nulo!");
+		}
 		try {
 			concesionariaService.delete(cocheId);
 			return ResponseEntity.noContent().build(); //204
